Update to tauri v2.0.0-beta.22
