+++ Dockerfile
@@ -1,19 +1,15 @@
-FROM rust:buster
+FROM ubuntu:22.04
 
 RUN apt-get update && apt-get install -y \
-    libwebkit2gtk-4.0-dev \
-    build-essential \
-    curl \
-    wget \
-    file \
-    libssl-dev \
-    libgtk-3-dev \
-    libayatana-appindicator3-dev\
-    librsvg2-dev\
-    libasound2-dev
+    libwebkit2gtk-4.1-dev \
+    libappindicator3-dev \
+    librsvg2-dev \
+    patchelf \
+    libasound2-dev \
+    nodejs \
+    npm
 
-RUN curl -sL https://deb.nodesource.com/setup_current.x | bash -
-RUN apt-get install -y nodejs
+RUN curl --proto '=https' --tlsv1.2 https://sh.rustup.rs -sSf | sh
 
 ADD . /app
 WORKDIR /app
