+++ docker-compose.yaml
@@ -1,4 +1,3 @@
-version: "3"
 services:
   build-linux:
     build:
@@ -11,4 +10,4 @@ services:
       TAURI_PRIVATE_KEY: ${TAURI_PRIVATE_KEY}
       TAURI_KEY_PASSWORD: ${TAURI_KEY_PASSWORD}
       APPLE_SIGNING_IDENTITY: ${APPLE_SIGNING_IDENTITY}
-    command: "npm run tauri build -- --target x86_64-unknown-linux-gnu --debug"
+    command: "npm run tauri build -- --target x86_64-unknown-linux-gnu"
