+++ src-tauri/Cargo.toml
@@ -1,33 +1,38 @@
 [package]
-name = "app"
+name = "pomodorolm"
 version = "0.1.0"
 description = "A Tauri App"
 authors = ["you"]
 license = ""
 repository = ""
-default-run = "app"
 edition = "2021"
 rust-version = "1.60"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "1.5.1", features = [] }
+tauri-build = { version = "2.0.0-beta", features = [] }
 
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.6.1", features = [ "app-all", "window-start-dragging", "fs-read-file", "system-tray", "shell-all", "icon-png", "devtools", "notification"] }
+tauri = { version = "2.0.0-beta", features = [ "devtools", "tray-icon", "image-png"] }
 image = { version = "0.25", features = ["png"] }
 tempfile = "3.10.1"
-tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
 rodio = "0.17.3"
 toml = "0.8.12"
-tokio = { version = "1.37.0", features = ["full"] }
+tokio = { version = "1", features = ["full"] }
 futures = "0.3.30"
-
+tauri-plugin-shell = "2.0.0-beta"
+tauri-plugin-notification = "2.0.0-beta"
+tauri-plugin-log = "2.0.0-beta"
+tokio-stream = "0.1.15"
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
 # If you use cargo directly instead of tauri's cli you can use this feature flag to switch between tauri's `dev` and `build` modes.
 # DO NOT REMOVE!!
 custom-protocol = [ "tauri/custom-protocol" ]
+
+[lib]
+name = "pomodorolm_lib"
+crate-type = ["staticlib", "cdylib", "lib"]
